# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseError = {
  code: string;
  message: string;
};

export interface ErrorResponse {
  error?: ErrorResponseError;
}

export interface RatingCategory {
  label: string;
  score: number;
}

export interface RatingDistribution {
  star: number;
  count: number;
  percent: number;
}

export interface Faq {
  question: string;
  answer: string;
}

export interface SocialProof {
  purchases: number;
  views: number;
  saves: number;
}

export interface ProductReview {
  id: string;
  userName: string;
  date: string;
  rating: number;
  text: string;
  pros?: string[];
  cons?: string[];
  helpful: number;
}

export interface BeforeAfterImage {
  before: string;
  after: string;
  label?: string;
}

export type ProductInventoryStatus = typeof ProductInventoryStatus[keyof typeof ProductInventoryStatus];


export const ProductInventoryStatus = {
  'in-stock': 'in-stock',
  'low-stock': 'low-stock',
  'out-of-stock': 'out-of-stock',
  'pre-order': 'pre-order',
} as const;

export interface Product {
  id: number;
  slug: string;
  name: string;
  description: string;
  businessId: string;
  businessName: string;
  businessVerified?: boolean;
  phone?: string;
  whatsapp?: string;
  city?: string;
  category: string;
  subcategory?: string;
  tags?: string[];
  price: number;
  originalPrice?: number;
  discountPercent?: number;
  currency: string;
  expiresAt?: string;
  isInstallmentAvailable: boolean;
  installmentMonths?: number;
  installmentProvider?: string;
  installmentDownPayment?: number;
  installmentMonthlyAmount?: number;
  rating: number;
  reviewCount: number;
  ratingBreakdown?: RatingCategory[];
  ratingDistribution?: RatingDistribution[];
  reviews?: ProductReview[];
  coverGradient: string;
  gallery?: string[];
  beforeAfterImages?: BeforeAfterImage[];
  inventoryStatus: ProductInventoryStatus;
  stockCount?: number;
  benefits?: string[];
  eligibleGroups?: string[];
  faqs?: Faq[];
  terms?: string;
  socialProof?: SocialProof;
  isFeatured: boolean;
  isNew: boolean;
  isPublished?: boolean;
  followerCount?: number;
  createdAt?: string;
  updatedAt?: string;
}

export type CreateProductBodyInventoryStatus = typeof CreateProductBodyInventoryStatus[keyof typeof CreateProductBodyInventoryStatus];


export const CreateProductBodyInventoryStatus = {
  'in-stock': 'in-stock',
  'low-stock': 'low-stock',
  'out-of-stock': 'out-of-stock',
  'pre-order': 'pre-order',
} as const;

export interface CreateProductBody {
  slug: string;
  name: string;
  description?: string;
  category: string;
  subcategory?: string;
  tags?: string[];
  price: number;
  originalPrice?: number;
  discountPercent?: number;
  currency?: string;
  isInstallmentAvailable?: boolean;
  installmentMonths?: number;
  installmentProvider?: string;
  installmentDownPayment?: number;
  installmentMonthlyAmount?: number;
  inventoryStatus?: CreateProductBodyInventoryStatus;
  stockCount?: number;
  benefits?: string[];
  eligibleGroups?: string[];
  faqs?: Faq[];
  terms?: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
  coverGradient?: string;
  isFeatured?: boolean;
  isNew?: boolean;
  isPublished?: boolean;
  followerCount?: number;
}

export type UpdateProductBodyInventoryStatus = typeof UpdateProductBodyInventoryStatus[keyof typeof UpdateProductBodyInventoryStatus];


export const UpdateProductBodyInventoryStatus = {
  'in-stock': 'in-stock',
  'low-stock': 'low-stock',
  'out-of-stock': 'out-of-stock',
  'pre-order': 'pre-order',
} as const;

/**
 * Partial update — all fields optional
 */
export interface UpdateProductBody {
  name?: string;
  description?: string;
  category?: string;
  subcategory?: string;
  tags?: string[];
  price?: number;
  originalPrice?: number;
  discountPercent?: number;
  isInstallmentAvailable?: boolean;
  installmentMonths?: number;
  installmentProvider?: string;
  installmentDownPayment?: number;
  installmentMonthlyAmount?: number;
  inventoryStatus?: UpdateProductBodyInventoryStatus;
  stockCount?: number;
  benefits?: string[];
  eligibleGroups?: string[];
  faqs?: Faq[];
  terms?: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
  coverGradient?: string;
  isFeatured?: boolean;
  isNew?: boolean;
  isPublished?: boolean;
  followerCount?: number;
}

export interface ProductMeta {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
}

export interface ProductListResponse {
  data: Product[];
  meta: ProductMeta;
}

export interface ProductSingleResponse {
  data: Product;
}

export type ListProductsParams = {
page?: number;
/**
 * @maximum 50
 */
per_page?: number;
/**
 * Full-text search query
 */
q?: string;
category?: string;
business_id?: string;
featured?: ListProductsFeatured;
sort?: ListProductsSort;
};

export type ListProductsFeatured = typeof ListProductsFeatured[keyof typeof ListProductsFeatured];


export const ListProductsFeatured = {
  true: 'true',
  false: 'false',
} as const;

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  created_at_desc: 'created_at_desc',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating_desc: 'rating_desc',
} as const;

export type ListBusinessProductsOwnerParams = {
page?: number;
/**
 * @maximum 50
 */
per_page?: number;
};