# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags:
        - health
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags:
        - products
      summary: List products (paginated, public)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
            maximum: 50
        - in: query
          name: q
          schema:
            type: string
          description: Full-text search query
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: business_id
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: string
            enum:
              - "true"
              - "false"
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - created_at_desc
              - price_asc
              - price_desc
              - rating_desc
            default: created_at_desc
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /products/{slug}:
    get:
      operationId: getProduct
      tags:
        - products
      summary: Get a single product by slug (public)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSingleResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /businesses/{businessId}/products/public:
    get:
      operationId: listBusinessProductsPublic
      tags:
        - products
      summary: List published products for a business (public, no auth)
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published products for the business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /businesses/{businessId}/products:
    get:
      operationId: listBusinessProductsOwner
      tags:
        - products
      summary: List all products for a business including drafts (owner-only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: All products for the business (including unpublished)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — token does not authorize this business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBusinessProduct
      tags:
        - products
      summary: Create a new product for a business (owner-only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSingleResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — token does not authorize this business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /businesses/{businessId}/products/{productId}:
    patch:
      operationId: updateBusinessProduct
      tags:
        - products
      summary: Update a product (owner-only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSingleResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — token does not authorize this business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBusinessProduct
      tags:
        - products
      summary: Delete a product (owner-only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: string
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Product deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — token does not authorize this business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message

    RatingCategory:
      type: object
      properties:
        label:
          type: string
        score:
          type: number
      required:
        - label
        - score

    RatingDistribution:
      type: object
      properties:
        star:
          type: integer
        count:
          type: integer
        percent:
          type: number
      required:
        - star
        - count
        - percent

    FAQ:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
        - question
        - answer

    SocialProof:
      type: object
      properties:
        purchases:
          type: integer
        views:
          type: integer
        saves:
          type: integer
      required:
        - purchases
        - views
        - saves

    ProductReview:
      type: object
      properties:
        id:
          type: string
        userName:
          type: string
        date:
          type: string
        rating:
          type: number
        text:
          type: string
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        helpful:
          type: integer
      required:
        - id
        - userName
        - date
        - rating
        - text
        - helpful

    BeforeAfterImage:
      type: object
      properties:
        before:
          type: string
        after:
          type: string
        label:
          type: string
      required:
        - before
        - after

    Product:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        businessId:
          type: string
        businessName:
          type: string
        businessVerified:
          type: boolean
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        category:
          type: string
        subcategory:
          type: string
        tags:
          type: array
          items:
            type: string
        price:
          type: integer
        originalPrice:
          type: integer
        discountPercent:
          type: integer
        currency:
          type: string
        expiresAt:
          type: string
          format: date-time
        isInstallmentAvailable:
          type: boolean
        installmentMonths:
          type: integer
        installmentProvider:
          type: string
        installmentDownPayment:
          type: integer
        installmentMonthlyAmount:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        ratingBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RatingCategory"
        ratingDistribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingDistribution"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ProductReview"
        coverGradient:
          type: string
        gallery:
          type: array
          items:
            type: string
        beforeAfterImages:
          type: array
          items:
            $ref: "#/components/schemas/BeforeAfterImage"
        inventoryStatus:
          type: string
          enum:
            - in-stock
            - low-stock
            - out-of-stock
            - pre-order
        stockCount:
          type: integer
        benefits:
          type: array
          items:
            type: string
        eligibleGroups:
          type: array
          items:
            type: string
        faqs:
          type: array
          items:
            $ref: "#/components/schemas/FAQ"
        terms:
          type: string
        socialProof:
          $ref: "#/components/schemas/SocialProof"
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        isPublished:
          type: boolean
        followerCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - description
        - businessId
        - businessName
        - category
        - price
        - currency
        - isInstallmentAvailable
        - rating
        - reviewCount
        - coverGradient
        - inventoryStatus
        - isFeatured
        - isNew

    CreateProductBody:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        tags:
          type: array
          items:
            type: string
        price:
          type: integer
        originalPrice:
          type: integer
        discountPercent:
          type: integer
        currency:
          type: string
        isInstallmentAvailable:
          type: boolean
        installmentMonths:
          type: integer
        installmentProvider:
          type: string
        installmentDownPayment:
          type: integer
        installmentMonthlyAmount:
          type: integer
        inventoryStatus:
          type: string
          enum:
            - in-stock
            - low-stock
            - out-of-stock
            - pre-order
        stockCount:
          type: integer
        benefits:
          type: array
          items:
            type: string
        eligibleGroups:
          type: array
          items:
            type: string
        faqs:
          type: array
          items:
            $ref: "#/components/schemas/FAQ"
        terms:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        coverGradient:
          type: string
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        isPublished:
          type: boolean
        followerCount:
          type: integer
      required:
        - slug
        - name
        - category
        - price

    UpdateProductBody:
      type: object
      description: Partial update — all fields optional
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        tags:
          type: array
          items:
            type: string
        price:
          type: integer
        originalPrice:
          type: integer
        discountPercent:
          type: integer
        isInstallmentAvailable:
          type: boolean
        installmentMonths:
          type: integer
        installmentProvider:
          type: string
        installmentDownPayment:
          type: integer
        installmentMonthlyAmount:
          type: integer
        inventoryStatus:
          type: string
          enum:
            - in-stock
            - low-stock
            - out-of-stock
            - pre-order
        stockCount:
          type: integer
        benefits:
          type: array
          items:
            type: string
        eligibleGroups:
          type: array
          items:
            type: string
        faqs:
          type: array
          items:
            $ref: "#/components/schemas/FAQ"
        terms:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        coverGradient:
          type: string
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        isPublished:
          type: boolean
        followerCount:
          type: integer

    ProductMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages

    ProductListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        meta:
          $ref: "#/components/schemas/ProductMeta"
      required:
        - data
        - meta

    ProductSingleResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Product"
      required:
        - data