# lib/api-zod/src/generated/types/productReview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductReview {
  id: string;
  userName: string;
  date: string;
  rating: number;
  text: string;
  pros?: string[];
  cons?: string[];
  helpful: number;
}